Fix calls to strncpy
Calls to strncpy don't guarantee that the destination buffer will be
terminated with a null-byte. GCC, with the -D_FORTIFY_SOURCE=2 flag, will
error out when strncpy is being called with a max size equal to the size of
the destination buffer.

Change these calls to pass in one less than the size of the destination
buffer, and explicitly add a null byte at the end of the buffer.